requirements.txt
streamlit
torch
torchvision
torchaudio
moviepy==1.0.3
faster-whisper
ctranslate2
openai
requests
numpy
pandas
pillow
opencv-python-headless
imageio[ffmpeg]
language-tool-python
psutil
tqdm
toml
packaging
watchdog
